Created argument parser and associated command line options
